Extracts server-side config tests into own test file.
